+++ src/Exception/HttpClientException.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Mailtrap\Exception;
 
+use JsonException;
 use Mailtrap\Helper\ResponseHelper;
 use Psr\Http\Message\ResponseInterface;
 
@@ -23,15 +24,18 @@ class HttpClientException extends HttpException
     {
         $errorMsg = '';
         $statusCode = $response->getStatusCode();
-        $body = ResponseHelper::toArray($response);
 
-        if (isset(self::ERROR_PREFIXES[$statusCode])) {
-            $errorMsg .= self::ERROR_PREFIXES[$statusCode] . ' Errors: ';
+        try {
+            $body = ResponseHelper::toArray($response);
+        } catch (JsonException|InvalidTypeException $e) {
+            $body['error'] = $response->getBody()->__toString();
         }
 
-        $errorMsg .= !empty($body['errors'])
-            ? (is_array($body['errors']) ? implode(' & ', $body['errors']) : $body['errors'])
-            : $body['error'];
+        if (isset(self::ERROR_PREFIXES[$statusCode])) {
+            $errorMsg .= self::ERROR_PREFIXES[$statusCode] . ' ';
+        }
+
+        $errorMsg .= trim('Errors: ' . self::getErrorMsg(!empty($body['errors']) ? $body['errors'] : $body['error']));
 
         return new self (
             !empty($errorMsg)
@@ -39,5 +43,39 @@ class HttpClientException extends HttpException
                 : sprintf('HTTP response code ("%d") received from the API server (no error info)', $statusCode),
             $statusCode
         );
+    }
+
+    /**
+     * It can be different structure of errors in the response...
+     *
+     * Examples:
+     * {"errors": ["'to' address is required", "'subject' is required"]}    400 errorS (array)
+     * {"error": "Incorrect API token"}                                     401 error  (string)
+     * {"errors": "Access forbidden"}                                       403 errorS (string)
+     * {"error": "Not found"}                                               404 error  (string)
+     * {"errors": {"name":["is too short (minimum is 2 characters)"]}}      422 errorS (array with key name)
+     *
+     *
+     * @param string|array $errors
+     *
+     * @return string
+     */
+    public static function getErrorMsg($errors): string
+    {
+        $errorMsg = '';
+        if (is_array($errors)) {
+            foreach ($errors as $key => $value) {
+                if (is_string($key)) {
+                    // add name of field
+                    $errorMsg .= $key . ' -> ';
+                }
+
+                $errorMsg .= self::getErrorMsg($value);
+            }
+        } else {
+            $errorMsg .= $errors . '. ';
+        }
+
+        return $errorMsg;
     }
 }