+++ src/Exception/InvalidTypeException.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Mailtrap\Exception;
+
+/**
+* Class InvalidTypeException
+*/
+class InvalidTypeException extends RuntimeException
+{
+}