+++ src/MailtrapSandboxClient.php
@@ -7,7 +7,8 @@ namespace Mailtrap;
 use Mailtrap\Api;
 
 /**
- * @method  Api\Sandbox\Emails   emails
+ * @method  Api\Sandbox\Emails     emails
+ * @method  Api\Sandbox\Project    projects
  *
  * Class MailtrapSandboxClient
  */
@@ -15,5 +16,6 @@ final class MailtrapSandboxClient extends AbstractMailtrapClient
 {
     public const API_MAPPING = [
         'emails' => Api\Sandbox\Emails::class,
+        'projects' => Api\Sandbox\Project::class,
     ];
 }