+++ tests/MailtrapTestCase.php
@@ -16,6 +16,9 @@ use Psr\Http\Client\ClientInterface;
 abstract class MailtrapTestCase extends TestCase
 {
     public const DEFAULT_API_KEY = 'some_api_key';
+    public const FAKE_ACCOUNT_ID = 10001;
+    public const FAKE_ACCOUNT_ACCESS_ID = 1000001;
+    public const FAKE_PROJECT_ID = 2436;
 
     protected function getHttpClientMock(): ClientInterface
     {
@@ -33,7 +36,7 @@ abstract class MailtrapTestCase extends TestCase
     protected function getConfigMock(): ConfigInterface
     {
         $config = $this->getMockBuilder(ConfigInterface::class)
-            ->onlyMethods(['getHttpClientBuilder', 'getApiToken', 'getHost'])
+            ->onlyMethods(['getHttpClientBuilder', 'getApiToken', 'getHost', 'isResponseThrowOnError'])
             ->getMock();
 
         $config
@@ -43,6 +46,10 @@ abstract class MailtrapTestCase extends TestCase
         $config
             ->method('getApiToken')
             ->willReturn(self::DEFAULT_API_KEY);
+
+        $config
+            ->method('isResponseThrowOnError')
+            ->willReturn(true);
 
         return $config;
     }